Add a regualr old Makefile
I enjoy using Makefiles. They, generally, work. They're quick. And they
have a low maintenance version. The Autoconf package has its purpose, if
you're using lots of packages that might be configured differently or
you offer variation, it can be useful.

I do /not/ enjoy CMake in any capacity. It is better than
the *combination* of Autoconf, Automake, and Libtool, but so is a mild
cold.

All we depend on is guile. It uses pkg-config, so calling pkg-config
from config.mk with the option to override should do everything we need.

Signed-off-by: Adam C. Emerson <azure@fox.blue>