Remove all of the Automake stuff
I don't mind Autoconf. I positively dislike Automake. I dislike the fact
that it often generates bad Makefiles. I dislike the fact that it ends
up willing the root of your project with files most of which end up
being empty. And I dislike that it is often unnecessary.

I just hate and despise Libtool. Thankfullythat's not an issue here.

Also the install target was kind of bad and the 'service' file is
outdated thanks to systemd. Yes, I like systemd. No, it's not because I
work for Red Hat.

Signed-off-by: Adam C. Emerson <azure@fox.blue>