Handle multiple hostnames and TLS properly
First, build our absolute URIs based on the incoming request rather than
a configured value. (More specifically, allow the user to supply a function
to extract parts of the URI from the request header.)

Second, redimension the cache to key on a set of request headers as well
as the method and path.

Signed-off-by: Adam C. Emerson <azure@fox.blue>